+++ V3D/FastFunctions.cpp
@@ -27,7 +27,7 @@ inline float fast_sqrt(float number)
 }
 
 template<class T>
-float fast_pow(T number, unsigned int n)
+float fast_pow(T number, size_t n)
 {
 	int result = 1;
 	while (n > 0) {
